Check that the BigBlueButton nginx definition exists

+++ bbb-install.sh
@@ -837,7 +837,9 @@ install_ssl() {
   if [ ! -f /etc/letsencrypt/live/$HOST/fullchain.pem ]; then
     rm -f /tmp/bigbluebutton.bak
     if ! grep -q $HOST /etc/nginx/sites-available/bigbluebutton; then  # make sure we can do the challenge
-      cp /etc/nginx/sites-available/bigbluebutton /tmp/bigbluebutton.bak
+      if [ -f /etc/nginx/sites-available/bigbluebutton ]; then
+        cp /etc/nginx/sites-available/bigbluebutton /tmp/bigbluebutton.bak
+      fi
       cat <<HERE > /etc/nginx/sites-available/bigbluebutton
 server_tokens off;
 server {
@@ -868,7 +870,6 @@ HERE
     if [ -z "$PROVIDED_CERTIFICATE" ]; then
       if ! certbot --email $EMAIL --agree-tos --rsa-key-size 4096 -w /var/www/bigbluebutton-default/ \
            -d $HOST --deploy-hook "systemctl restart nginx" $LETS_ENCRYPT_OPTIONS certonly; then
-        cp /tmp/bigbluebutton.bak /etc/nginx/sites-available/bigbluebutton
         systemctl restart nginx
         err "Let's Encrypt SSL request for $HOST did not succeed - exiting"
       fi
